model/requirements.txt
# Flask & Server
Flask==2.2.5
flask-cors==3.0.10
gunicorn==23.0.0
Werkzeug==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
click==8.1.8

# MongoDB
Flask-PyMongo==2.3.0
pymongo==4.13.2
dnspython==2.6.1
bson==0.5.10

# Deep Learning & Image Processing
deepface==0.0.93
opencv-contrib-python-headless==4.11.0.86
opencv-python-headless==4.11.0.86
mtcnn==1.0.0
retina-face==0.0.17
tensorflow==2.19.0
tf_keras==2.19.0
keras==3.10.0
torch==2.7.1+cpu
numpy==1.26.4
scipy==1.16.0
scikit-learn==1.7.0

# Cloudinary
cloudinary==1.41.0

# Utilities
requests==2.32.4
Pillow==10.2.0
tqdm==4.67.1
python-dotenv==1.1.1
protobuf==4.25.8
h5py==3.14.0

# Optional 
ffmpeg==1.4
imageio==2.37.0
imageio-ffmpeg==0.6.0

gunicorn
